finished work on ullrs trueshot CDR module
