+++ src/Parser/MarksmanshipHunter/CHANGELOG.js
@@ -1,4 +1,5 @@
 export default `
+16-10-2017 - Added Ullrs Trueshot CDR support. (by Putro)
 14-10-2017 - Added Focus Capped Statistic Box (By leapis)
 08-10-2017 - Added FocusTracker Module and FocusTracker Graph module and additional supporting modules (By leapis)
 05-10-2017 - remove Cyclonic Burst from Cooldown view. (by Putro)
