+++ src/Parser/MarksmanshipHunter/CombatLogParser.js
@@ -26,6 +26,9 @@ import Tier20_4p from './Modules/Items/Tier20_4p';
 //Spells
 import Trueshot from './Modules/Spells/Trueshot';
 
+//Items
+import UllrsFeatherSnowshoes from "./Modules/Items/UllrsFeatherSnowshoes";
+
 
 class CombatLogParser extends CoreCombatLogParser {
 
@@ -47,6 +50,7 @@ class CombatLogParser extends CoreCombatLogParser {
     //Items
     tier20_2p: Tier20_2p,
     tier20_4p: Tier20_4p,
+    ullrsFeatherSnowshoes: UllrsFeatherSnowshoes,
 
     //Spells
     trueshot: Trueshot,
